Move mixin to a helper module
